[TimeSeries] Supporting By parameter.
 - new structure
 - TimeSeriesByCategory

TODO:
 - features_importance
 - deploySQL
 - regression_report
 - score
 - plot
